dash/requirements.txt
dash == 2.0.0
dash-daq == 0.5.0
dash-bootstrap-components == 1.0.2
plotly-express==0.4.1